Added and refactored the UI Elements and their interfaces defined in the js driver package<.
Implemented MutationRecords which are an append-only store of the UI mutation data.
It should allow to retrieve the state of the UI based on the mutations stored in chornological order.
Rationale being that UI state is in general path dependent due to non-commutatitvity.

Exple: mutating an element property is only possible after it has been created.
Cannot retrieve state if there is not at least a partial order of mutations.

Started rewriting the Persistence functions that hsould be used to store properties in the js Session/Localstore.
Remains to create the logic of setting a persistence store option on for an Element.