Add multiple important fix and  changes
Initially the UI state was using event sourcing to record UI mutations.
It is wasteful in terms of resource and possibly untractable as it is.
Issued changes to make UI state directly translate into UI change by application of
multiple independent, idempotent, thus commutative functions.
Therefore the idea of purity (UI as a pure function of application state)
Here we have more granularity allowing for surgical updates of the UI without re rendering.

Now the mutation event has access to the old value, pre mutation.

New types of *ui.Element :
- ui.BasicElement which expose a simple interface used to build
UI trees. (alongside the ViewElements which are a second type of object not too dissimilar
but allowing for named version of a *ui.Element)
- ui.Observable which exposes an interface for watachable objects by *ui.Element