Add a few fix to the native event handling, the Mutation Observing
There was a bug in the handling of  Unwatch where the mutation event handlers
would remain. Since observing an element porperties is solely registered
via the category, property and observer Element ID, when an observee was
recreated with the same id, the old mutation event handlers would fire.

Essentially, causing the possibility of multiple registration.