Make router accessible for use by mounted ViewElements and modifies View activation
One thing that was not well-thought is the issue of how deep-linking should be done.
It requires for components (*Element (sub)trees) to be able to create links.
But links are created by the router which the App tree remains unaware of (because we don't want to couple the tree and the router)
Hence, to simplify things, once a new router is created, it becomes the de facto only router and is stored as a singleton.
It means that2 projects may not share the same library: each workspace ought to retain its own version of the particleui library.

It is how it ought ot be anyway has we intend to have the same behaviour to handle changes tocss files (styles may not be shared between different project automatically)

Also, view activation has been reworked.
For now, testing involves real -app testing but it has not be done at the moment of this commit.