Adds lifecyclehandlers for UI tree root function
Reworks data loading:
- based on the state transition API

Reworks the mutation replay and mutation capture logic (wip)

On successful replay, the replay buffer persisted value should not be cleared.
It should only be cleared if replay was erroneous for some reason.
And even then, maybe we will want to implement some sort of backup to avoid
erasing all data due to a transient system failure.