Implemented a working HTML backend for server side page generation that is fully compatible with the syscall/js API for isomorphic rendering between server and client. In a nutshell, we should now be able to run the same code on the client and on the server.
Other small (and not so small) changes:

* added Length method to List ui.Value.
* modified the signature of ui.DoSync : it now takes a *ui.Element argument
in order to have access to the router for the document that the element is a part of.
Indeed we should be able to have access to a main UI thread/goroutine per document.
Especially relevant on the server where we may generate multiple documents cocnurrently.
On the client, there is only one main UI thread in general so it might be less relevant.
* rewrote the projectlauncher: new file structure

TODOS:

* apply changes to ssr and ssg modes (dom ssr/ssg files) similarly to csr where applicable
* test stylesheet rendering?
* prettify/beautify/format html output before writing to file